requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autopep8==2.1.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.2.0
control==0.9.4
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dgl==2.1.0
dglgo==0.0.1
docutils==0.20.1
easydict==1.13
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.4.0
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
imageio==2.34.0
imagesize==1.4.1
importlib_metadata==7.0.2
ipdb==0.13.13
ipykernel==6.29.3
ipython==8.22.2
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nose==1.3.7
notebook==7.1.1
notebook_shim==0.2.4
numba==0.59.0
numpy==1.26.3
numpydoc==1.6.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.42
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyemd==1.0.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rdkit==2023.9.5
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
torch==2.2.0+cu121
torch_geometric==2.5.1
torch_sparse==0.6.18
torchaudio==2.2.0+cu121
torchdata==0.7.1
torchvision==0.17.0+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
triton==2.2.0
typer==0.9.0
tyro>=0.8.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.8.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.42.0
yarl==1.9.4
zipp==3.17.0
